+++ core/testsuite/src/main/java/com/blazebit/persistence/testsuite/entity/BookEntity.java
@@ -1,0 +1,51 @@
+/*
+ * Copyright 2014 - 2018 Blazebit.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.blazebit.persistence.testsuite.entity;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Table;
+import java.io.Serializable;
+
+/**
+ *
+ * @author Jan-Willem Gmelig Meyling
+ * @since 1.2.1
+ */
+@Entity
+@Table(name = "book")
+public class BookEntity extends Ownable implements Serializable {
+    private static final long serialVersionUID = 1L;
+
+    private String isbn;
+
+    public BookEntity() {
+    }
+
+    public BookEntity(Long id) {
+        super(id);
+    }
+
+    @Column(unique = true, name = "isbn")
+    public String getIsbn() {
+        return isbn;
+    }
+
+    public void setIsbn(String isbn) {
+        this.isbn = isbn;
+    }
+}